+++ FilmManager/FilmManagerCore/Filters/FilmFilters.cs
@@ -8,7 +8,7 @@ namespace FilmManagerCore.Filters
 {
     public class FilmFilters
     {
-        public int? GenreId { get; set; }
+        public long? GenreId { get; set; }
         public int? Year { get; set; }
         public float? Rating { get; set; }
         public int? SelfRating { get; set; }
